chore(team): split mailbox helpers and tests
