+++ src/team/manager/mailbox_worker.rs
@@ -1,0 +1,41 @@
+use std::sync::Arc;
+use std::time::Duration;
+
+use super::{TeamManager, TeamRemoteRelayWorkerSettings};
+
+impl TeamManager {
+    pub fn spawn_remote_relay_worker(self: Arc<Self>, settings: TeamRemoteRelayWorkerSettings) {
+        tokio::spawn(async move {
+            let mut ticker = tokio::time::interval(Duration::from_secs(
+                settings.poll_interval_secs.max(1) as u64,
+            ));
+            ticker.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
+            loop {
+                ticker.tick().await;
+                match self
+                    .relay_remote_messages_once(
+                        settings.batch_limit,
+                        settings.max_attempts,
+                        settings.retry_delay_secs,
+                    )
+                    .await
+                {
+                    Ok(summary) => {
+                        if summary.scanned > 0 {
+                            tracing::debug!(
+                                scanned = summary.scanned,
+                                delivered = summary.delivered,
+                                retried = summary.retried,
+                                dead_lettered = summary.dead_lettered,
+                                "team relay worker tick"
+                            );
+                        }
+                    }
+                    Err(err) => {
+                        tracing::warn!("team relay worker tick failed: {}", err);
+                    }
+                }
+            }
+        });
+    }
+}
